Changed deletion logic to use indexes

+++ Services/TasksTrackerService/EmployeesTasksTracker.TasksTrackerService.Application/Handlers/Projects/DeleteProjectHandler.cs
@@ -8,27 +8,19 @@ namespace EmployeesTasksTracker.TasksTrackerService.Application.Handlers.Project
     public class DeleteProjectHandler : IRequestHandler<DeleteProjectCommand, bool>
     {
         private readonly IProjectsRepo _projectsRepo;
-        private readonly ITasksRepo _tasksRepo;
 
-        public DeleteProjectHandler(IProjectsRepo projectsRepo, ITasksRepo tasksRepo)
+        public DeleteProjectHandler(IProjectsRepo projectsRepo)
         {
             _projectsRepo = projectsRepo;
-            _tasksRepo = tasksRepo;
         }
 
         public async Task<bool> Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
         {
-            var relatedTasks = await _tasksRepo.GetAllFilteredAsync(default, default, request.Id, cancellationToken);
+            var canDelete = await _projectsRepo.CheckDeletionCapability(request.Id, cancellationToken);
 
-            if (relatedTasks.Any())
+            if (!canDelete)
             {
-                foreach (var task in relatedTasks)
-                {
-                    if (task.Status != Core.Enums.Status.Canceled || task.Status != Core.Enums.Status.Completed)
-                    {
-                        throw new DomainException($"Couldn't delete project! Task {task.Name} is {task.Status}!");
-                    }
-                }
+                throw new DomainException($"Couldn't delete project! Task isn't completed or cancelled!");
             }
 
             return await _projectsRepo.DeleteAsync(request.Id, cancellationToken);
